requirements.txt
kafka-python==2.0.2
pyspark==3.5.0
scikit-learn==1.5.0
pandas==2.2.0
numpy==1.26.0
streamlit==1.26.0
joblib==1.3.2
requests==2.31.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0 